% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HallPolynomials.R
\name{HallPolynomials}
\alias{HallPolynomials}
\title{Hall polynomials}
\usage{
HallPolynomials(mu, nu)
}
\arguments{
\item{mu, nu}{integer partitions}
}
\value{
A list of lists. Each of these lists has two elements: an integer
  partition \eqn{\lambda} in the field \code{lambda}, and a univariate
  \code{qspray} polynomial in the field \code{polynomial}, the Hall
  polynomial \eqn{g^{\lambda}_{\mu,\nu}(t)}. Every coefficient of a
  Hall polynomial is an integer.
}
\description{
Hall polynomials \eqn{g^{\lambda}_{\mu,\nu}(t)} for given
  integer partitions \eqn{\mu} and \eqn{\nu}.
}
\note{
This function is slow.
}
\examples{
HallPolynomials(c(2, 1), c(1, 1))
}
