% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPP.R
\name{JackPol}
\alias{JackPol}
\title{Jack polynomial - C++ implementation}
\usage{
JackPol(n, lambda, alpha, which = "J")
}
\arguments{
\item{n}{number of variables, a positive integer}

\item{lambda}{an integer partition, given as a vector of decreasing
integers}

\item{alpha}{rational number, given as a string such as
\code{"2/3"} or as a \code{bigq} number}

\item{which}{which Jack polynomial, \code{"J"}, \code{"P"}, \code{"Q"},
or \code{"C"}}
}
\value{
A \code{qspray} multivariate polynomial.
}
\description{
Returns a Jack polynomial.
}
\examples{
JackPol(3, lambda = c(3, 1), alpha = "2/5")
}
