% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JackSymPol.R
\name{JackSymPol}
\alias{JackSymPol}
\title{Jack polynomial with symbolic Jack parameter}
\usage{
JackSymPol(n, lambda, which = "J")
}
\arguments{
\item{n}{number of variables, a positive integer}

\item{lambda}{an integer partition, given as a vector of decreasing
integers}

\item{which}{which Jack polynomial, \code{"J"}, \code{"P"}, \code{"Q"},
or \code{"C"}}
}
\value{
A \code{symbolicQspray} object.
}
\description{
Returns the Jack polynomial with a symbolic Jack parameter.
}
\examples{
JackSymPol(3, lambda = c(3, 1))
}
