% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiedMacdonaldPolynomial.R
\name{modifiedMacdonaldPol}
\alias{modifiedMacdonaldPol}
\title{Modified Macdonald polynomial}
\usage{
modifiedMacdonaldPol(n, mu)
}
\arguments{
\item{n}{number of variables, a positive integer}

\item{mu}{integer partition}
}
\value{
A \code{symbolicQspray} multivariate polynomial, the modified
  Macdonald polynomial associated to the integer partition \code{mu}.
  It has two parameters and its coefficients are polynomials in these
  parameters.
}
\description{
Returns the modified Macdonald polynomial associated to a
  given integer partition.
}
