% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtKostkaPolynomials.R
\name{qtKostkaPolynomials}
\alias{qtKostkaPolynomials}
\title{qt-Kostka polynomials}
\usage{
qtKostkaPolynomials(mu)
}
\arguments{
\item{mu}{integer partition}
}
\value{
A list. The qt-Kostka polynomials are usually denoted by
  \eqn{K_{\lambda, \mu}(q, t)} where \eqn{q} and \eqn{t} denote the two
  variables and \eqn{\lambda} and \eqn{\mu} are two integer partitions.
  One obtains the Kostka-Foulkes polynomials by substituting \eqn{q}
  with \eqn{0}.
  For a given partition \eqn{\mu}, the function returns the
  polynomials \eqn{K_{\lambda, \mu}(q, t)} as \code{qspray} objects
  for all partitions \eqn{\lambda} of the same weight as \eqn{\mu}. The
  generated list is a list of lists with two elements: the integer
  partition \eqn{\lambda} and the polynomial.
}
\description{
qt-Kostka polynomials, aka Kostka-Macdonald polynomials.
}
