% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkewJackPol.R
\name{SkewJackSymPol}
\alias{SkewJackSymPol}
\title{Skew Jack polynomial with symbolic Jack parameter}
\usage{
SkewJackSymPol(n, lambda, mu, which = "J")
}
\arguments{
\item{n}{positive integer, the number of variables}

\item{lambda}{outer integer partition of the skew partition}

\item{mu}{inner integer partition of the skew partition; it must be a
subpartition of \code{lambda}}

\item{which}{which skew Jack polynomial, \code{"J"}, \code{"P"}, \code{"Q"}
or \code{"C"}}
}
\value{
A \code{symbolicQspray} polynomial.
}
\description{
Computes a skew Jack polynomial with a symbolic Jack parameter.
}
\examples{
SkewJackSymPol(3, c(3,1), c(2))
}
