% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JackSymbolicCombination.R
\name{symbolicJackCombination}
\alias{symbolicJackCombination}
\title{Symmetric polynomial in terms of symbolic Jack polynomials}
\usage{
symbolicJackCombination(qspray, which = "J", check = TRUE)
}
\arguments{
\item{qspray}{a \code{qspray} object or a \code{symbolicQspray} object
defining a symmetric polynomial}

\item{which}{which Jack polynomials, \code{"J"}, \code{"P"}, \code{"Q"} or
\code{"C"}}

\item{check}{Boolean, whether to check the symmetry}
}
\value{
A list defining the combination. Each element of this list is a
  list with two elements: \code{coeff}, a \code{bigq} number, and
  \code{lambda}, an integer partition; then this list corresponds to the
  term \code{coeff * JackSymPol(n, lambda, which)}, where \code{n} is
  the number of variables in the symmetric polynomial.
}
\description{
Expression of a symmetric polynomial as a linear combination
  of Jack polynomials with a symbolic Jack parameter.
}
