% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kostka.R
\name{symbolicKostkaJackNumbersWithGivenLambda}
\alias{symbolicKostkaJackNumbersWithGivenLambda}
\title{Kostka-Jack numbers with symbolic Jack parameter for a
  given \eqn{\lambda}}
\usage{
symbolicKostkaJackNumbersWithGivenLambda(lambda)
}
\arguments{
\item{lambda}{integer partition}
}
\value{
A named list of \code{ratioOfQsprays} objects. The elements of this
  list are the Kostka-Jack numbers \eqn{K_{\lambda,\mu}(\alpha)} and
  its names correspond to the partitions \eqn{\mu}.
}
\description{
Kostka-Jack numbers \eqn{K_{\lambda,\mu}(\alpha)} with a
  symbolic Jack parameter \eqn{\alpha} for a given
  integer partition \eqn{\lambda}.
}
\examples{
symbolicKostkaJackNumbersWithGivenLambda(c(3, 1))
}
\seealso{
\code{\link{KostkaJackNumbersWithGivenLambda}},
  \code{\link{symbolicKostkaJackNumbers}}.
}
