% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPP.R
\name{Zonal}
\alias{Zonal}
\title{Evaluation of zonal polynomial - C++ implementation}
\usage{
Zonal(x, lambda)
}
\arguments{
\item{x}{values of the variables, a vector of \code{bigq} numbers, or a
vector that can be coerced as such (e.g. \code{c("2", "5/3")})}

\item{lambda}{an integer partition, given as a vector of decreasing
integers}
}
\value{
A \code{bigq} number.
}
\description{
Evaluates a zonal polynomial. The zonal polynomials are the
  Jack \eqn{C}-polynomials with Jack parameter \eqn{\alpha=Z}.
}
\examples{
Zonal(c("1", "3/2", "-2/3"), lambda = c(3, 1))
}
