% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPP.R
\name{ZonalPol}
\alias{ZonalPol}
\title{Zonal polynomial - C++ implementation}
\usage{
ZonalPol(n, lambda)
}
\arguments{
\item{n}{number of variables, a positive integer}

\item{lambda}{an integer partition, given as a vector of decreasing
integers}
}
\value{
A \code{qspray} multivariate polynomial.
}
\description{
Returns a zonal polynomial. The zonal polynomials are the
  Jack \eqn{C}-polynomials with Jack parameter \eqn{\alpha=Z}.
}
\examples{
( zonal <- ZonalPol(3, lambda = c(3, 1)) )
zonal == JackPol(3, lambda = c(3, 1), alpha = "2", which = "C")
}
