% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPP.R
\name{ZonalQ}
\alias{ZonalQ}
\title{Evaluation of zonal quaternionic polynomial - C++ implementation}
\usage{
ZonalQ(x, lambda)
}
\arguments{
\item{x}{values of the variables, a vector of \code{bigq} numbers, or a
vector that can be coerced as such (e.g. \code{c("2", "5/3")})}

\item{lambda}{an integer partition, given as a vector of decreasing
integers}
}
\value{
A \code{bigq} number.
}
\description{
Evaluates a zonal quaternionic polynomial. The quaternionic
  zonal polynomials are the Jack \eqn{C}-polynomials with Jack
  parameter \eqn{\alpha=1/Z}.
}
\examples{
ZonalQ(c("1", "3/2", "-2/3"), lambda = c(3, 1))
}
